+++ autochain/examples/get_weather_with_openai_function_agent.py
@@ -1,6 +1,6 @@
 import json
 
-from autochain.agent.openai_funtions_agent.openai_functions_agent import (
+from autochain.agent.openai_functions_agent.openai_functions_agent import (
     OpenAIFunctionsAgent,
 )
 from autochain.chain.chain import Chain
