+++ autochain/workflows_evaluation/openai_function_agent_eval/get_weather_test.py
@@ -1,6 +1,6 @@
 import json
 
-from autochain.agent.openai_funtions_agent.openai_functions_agent import (
+from autochain.agent.openai_functions_agent.openai_functions_agent import (
     OpenAIFunctionsAgent,
 )
 from autochain.models.chat_openai import ChatOpenAI
