+++ docs/examples.md
@@ -182,7 +182,7 @@ Here is one example of it to get weather information.
 import json
 import logging
 
-from autochain.agent.openai_funtions_agent.openai_functions_agent import (
+from autochain.agent.openai_functions_agent.openai_functions_agent import (
     OpenAIFunctionsAgent,
 )
 from autochain.chain.chain import Chain
